wrap transports in the same way as registry library
